Reduce the fit image size to speed up the board booting
The original initramfs image.fit file contain many packages, so the size
was quite large, this is mainly for the purpose of debugging.
Currently, since the jh7110 functions have now stabilized, it is possible
to reduce the size of the image in order to speed up the boot process.

We also still keep the original buildroot initramfs config as the debug:
conf/buildroot_initramfs_config_debug
How to use the config to generate the big image.fit:
$ HWBOARD_CONFIG=debug make -j(nproc)

Note that do not use the debug config to gengerate the sdcard.img,
for it is no sense.

Signed-off-by: Andy Hu <andy.hu@starfivetech.com>